requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
Django==3.2.6
fastapi==0.68.0
Flask==2.0.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
image==1.5.33
imageio==2.9.0
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
logging==0.4.9.6
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
networkx==2.6.2
numpy==1.21.2
opencv-python==4.5.3.56
pandas==1.3.2
parso==0.8.2
PassportEye==2.1.0
pdfminer==20191125
pexpect==4.8.0
phonenumbers==8.12.29
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycryptodome==3.10.1
pydantic==1.8.2
Pygments==2.9.0
pyodbc==4.0.31
pyparsing==2.4.7
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
pyzmq==22.1.0
requests==2.26.0
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.22
sqlparse==0.4.1
starlette==0.14.2
tesseract==0.1.3
threadpoolctl==2.2.0
tifffile==2021.8.8
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
wcwidth==0.2.5
Werkzeug==2.0.1
